+++ R/class_summary.R
@@ -59,7 +59,7 @@
 #' @examples
 #' data(galap)
 #' #fit a multimodel SAR and get the model table
-#' mf <- sar_average(data = galap)
+#' mf <- sar_average(data = galap, grid_start = "none")
 #' summary(mf)
 #' summary(mf)$Model_table
 #' #Get a summary of the fit of the linear power model