+++ inst/model_factory.R
@@ -33,9 +33,9 @@ cat_roxygen <- function(model, funName, fileName){
   cat1("\n")
   cat1(paste0("#' @description ","Fit the ", model$name,
               " model to SAR data.", "\n"))
-  cat1(paste0("#' @usage ", funName,
-              "(data, start = NULL, grid_start = FALSE, grid_n = NULL,", 
-              " normaTest = 'none',", "\n"))
+  cat1(paste0("#' @usage ", funName, "(data, start = NULL,",
+              " grid_start = 'partial',", "\n"))
+  cat1(paste0("#'   grid_n = NULL, normaTest = 'none',", "\n"))
   cat1(paste0("#'   homoTest = 'none', homoCor = 'spearman')\n"))
   cat1(paste0("#' @param ", "data ", "A dataset in the form of a dataframe ",
                 "with two columns: \n"))
@@ -44,13 +44,12 @@ cat_roxygen <- function(model, funName, fileName){
   cat1(paste0("#'   of each island/site.\n"))
   cat1(paste0("#' @param ", "start ", "NULL or custom parameter start", 
               " values for the optimisation algorithm.\n"))
-  cat1(paste0("#' @param ", "grid_start ", "Logical argument specifying whether", 
-              " a grid search procedure should be implemented to test multiple",
-              " starting parameter values. The default is set to FALSE, but",
-              " for certain models (e.g. Gompertz, Chapman Richards), we",
-              " advice using it to ensure an optimal fit.\n"))
-  cat1(paste0("#' @param ", "grid_n ", "If \\code{grid_start = TRUE}, the",
-              " number of points sampled in the model parameter space.\n"))
+  cat1(paste0("#' @param ", "grid_start ", "Should a grid search procedure be", 
+              " implemented to test multiple starting parameter values.",
+              " Can be one of 'none', 'partial' or 'exhaustive' The default",
+              " is set to 'partial'.\n"))
+  cat1(paste0("#' @param ", "grid_n ", "If \\code{grid_start = exhaustive}, the",
+              " number of points sampled in the starting parameter space.\n"))
   cat1(paste0("#' @param ", "normaTest ", "The test used to test the", 
               " normality of the residuals of the\n"))
   cat1(paste0("#'   model. Can be any of 'lillie' (Lilliefors ", 
@@ -95,7 +94,11 @@ cat_roxygen <- function(model, funName, fileName){
               "\n\n")) 
   cat1(paste0("#'   A selection of information criteria (e.g. AIC, BIC) are", 
               " returned and can be used to compare models\n")) 
-  cat1(paste0("#'   (see also \\code{\\link{sar_average}})\n")) 
+  cat1(paste0("#'   (see also \\code{\\link{sar_average}}).\n")) 
+  cat1(paste0("#'   \n")) 
+  cat1(paste0("#'   As grid_start has a random component, when",
+              " \\code{grid_start != 'none'} in your model fitting, you can\n")) 
+  cat1(paste0("#'    get slightly different results each time you fit a model\n")) 
   cat1(paste0("#' @importFrom ", "stats lm quantile\n")) 
   
   cat1(paste0("#' @return ", "A list of class 'sars' with the following", 
@@ -146,10 +149,21 @@ cat_roxygen <- function(model, funName, fileName){
   cat1(paste0("#'   relationship: biology and statistics. Journal of", 
               " Biogeography, 39, 215-231.\n"))
   
-  cat1(paste0("#' @examples", "\n", "#' data(galap)", "\n",
-              "#' fit <- ", funName,
-              "(galap)", "\n", "#' summary(fit)","\n", "#' plot(fit)\n"))
-  
+  #grid_start makes example too long for betap, so just for this model turn
+  #grid_start off
+  if (funName == "sar_betap"){
+    cat1(paste0("#' @examples", "\n", 
+                "#' #Grid_start turned off for speed (not recommended)", "\n",
+                "#' data(galap)", "\n",
+                "#' fit <- ", funName,
+                "(galap, grid_start = 'none')", "\n", 
+                "#' summary(fit)","\n", "#' plot(fit)\n"))
+  } else{
+    cat1(paste0("#' @examples", "\n", "#' data(galap)", "\n",
+                "#' fit <- ", funName,
+                "(galap)", "\n", "#' summary(fit)","\n", "#' plot(fit)\n"))
+  }
+
   cat1(paste0("#' @export"))
   
 }#eo cat_roxygen
@@ -191,9 +205,10 @@ model_factory <- function(f, overwrite = FALSE){
   cat1("\n")
   
   #function definition
-  cat1(paste0(funName,' <- function(data, start = NULL, grid_start = FALSE,', 
-                        ' \ngrid_n = NULL, normaTest =  "none", homoTest =',
-                        ' "none", homoCor = "spearman"){',"\n"))
+  cat1(paste0(funName,' <- function(data, start = NULL,', 
+                        ' \ngrid_start = "partial", grid_n = NULL,', 
+                        ' \nnormaTest =  "none", homoTest =',
+                        ' \n"none", homoCor = "spearman"){',"\n"))
   
   #checks
   cat1("if (!(is.matrix(data) | is.data.frame(data)))", 
@@ -208,10 +223,12 @@ model_factory <- function(f, overwrite = FALSE){
   cat1("homoCor <- match.arg(homoCor, c('spearman', 'pearson',\n")
   cat1("'kendall'))","\n")
   cat1("}\n")
-  cat1("if (!is.logical(grid_start)) stop('grid_start should be logical')\n")
-  cat1("if (grid_start){\n")
+  cat1("if (!(grid_start %in% c('none', 'partial', 'exhaustive'))){\n")
+  cat1("stop('grid_start should be one of none, partial or exhaustive')\n")
+  cat1("}\n")
+  cat1("if (grid_start == 'exhaustive'){\n")
   cat1("  if (!is.numeric(grid_n))\n")
-  cat1("  stop('grid_n should be numeric if grid_start == TRUE')\n")
+  cat1("  stop('grid_n should be numeric if grid_start == exhaustive')\n")
   cat1("  }\n")
   
   #data ordering and column naming (assuming Area then Species Richness)