+++ tests/testthat/test_multi.R
@@ -2,8 +2,10 @@ context("sar_average")
 
 test_that("sar_average returns correct results", {
 
-  fit3 <- sar_average(data = galap)
-  expect_equal(round(sum(fit3$mmi), 1), 1643.5)
+  fit2 <- sar_average(data = galap, grid_start = "none")
+  expect_equal(round(sum(fit2$mmi), 1), 1643.5)
+  fit3 <- sar_average(data = galap)#grid start on so round to 0
+  expect_equal(round(sum(fit3$mmi), 0), 1643)
   expect_output(str(fit3), "List of 2")
   expect_is(fit3, "multi")
   expect_match(fit3$details$homo_test, "none")
@@ -14,7 +16,7 @@ test_that("sar_average returns correct results", {
   fit4 <- sar_average(data = galap, normaTest = "lillie", 
                       homoTest = "cor.fitted",
                       neg_check = FALSE)
-  expect_equal(round(sum(fit4$mmi), 1), 1633.9)
+  expect_equal(round(sum(fit4$mmi), 0), 1634)#grid_start on so round to 0
   expect_match(fit4$details$norm_test, "lillie")
   expect_match(fit4$details$homo_test, "cor.fitted")
   expect_equal(length(fit4$details$mod_names), 14)
@@ -23,14 +25,13 @@ test_that("sar_average returns correct results", {
                            homoCor = "correlation"))
 })
 
-
 test_that("sar_average using fit_collection object works", {
   ff <- sar_multi(data = galap, obj = c("power", "p1", "loga", "monod",
                                         "linear"))
   expect_warning(sar_average(obj = ff, data = galap, normaTest = "none",
                     neg_check = FALSE, homoTest = "cor.fitted"))
-  fit3 <- sar_average(obj = ff, data = galap)
-  expect_equal(round(sum(fit3$mmi), 1), 1662.8)
+  fit3 <- sar_average(obj = ff, data = galap)#grid start on so round to 0
+  expect_equal(round(sum(fit3$mmi), 0), 1663)
   expect_output(str(fit3), "List of 2")
   expect_is(fit3, "multi")
   expect_match(fit3$details$homo_test, "none")
@@ -38,11 +39,10 @@ test_that("sar_average using fit_collection object works", {
   expect_error(sar_multi(5), "argument is of length zero")
 })
 
-
-
 test_that("confidence intervals are correct", {
   skip_on_cran()
-  fit3 <- expect_warning(sar_average(data = galap, normaTest = "none", homoTest = "none",
+  fit3 <- expect_warning(sar_average(data = galap, grid_start = "none",
+                                     normaTest = "none", homoTest = "none",
                     neg_check = FALSE, confInt = TRUE, ciN = 20))
   ci <- fit3$details$confInt
   expect_equal(nrow(ci), nrow(galap))
@@ -54,7 +54,8 @@ test_that("confidence intervals are correct", {
 
 test_that("confidence intervals works with AICc", {
   skip_on_cran()
-  fit3 <- expect_warning(sar_average(data = galap, normaTest = "none", homoTest = "none",
+  fit3 <- expect_warning(sar_average(data = galap, grid_start = "none",
+                                     normaTest = "none", homoTest = "none",
                       neg_check = FALSE, confInt = TRUE, crit = "AICc", ciN = 20))
   ci <- fit3$details$confInt
   expect_equal(nrow(ci), nrow(galap))