+++ tests/testthat/test_power.R
@@ -17,7 +17,7 @@ test_that("sar_power returns correct results", {
 #changed neg_expo function allowing z to be > 1, and also the change to 
 #asymp setting z to Rplus rather than R
 test_that("neg_expo and asymp returns correct results", {
-  fit <- sar_negexpo(niering)
+  fit <- sar_negexpo(niering, grid_start = "none")
   expect_equal(round(fit$AICc, 2), 207.39)
   expect_equal(as.vector(round(fit$par[2], 2)), 26.14)
   fit2 <- sar_asymp(niering)
@@ -29,7 +29,7 @@ test_that("neg_expo and asymp returns correct results", {
 test_that("sar_power summary returns correct results", {
   fit <- sar_power(galap, normaTest = "lillie")
   fs <- summary(fit)
-  expect_equal(round(sum(fs$residuals), 1), -31.1)
+  expect_equal(round(sum(fs$residuals), 0), -31)#grid start so round to 0
   expect_output(str(fs), "List of 16")
   expect_is(fs, "summary.sars")
   expect_equal(round(fs$normaTest[[2]]$p.value, 3), 0.056)